+++ apps/payment-service/src/main/java/com/server/payment/domain/IPaymentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IPaymentService {
     List<PaymentEntity> fetch();
     PaymentEntity create(PaymentEntity payment);
-    PaymentEntity getById(Long id);
-    PaymentEntity updateStatusById(Long id, Status paymentStatus);
-    void deleteById(Long id);
+    PaymentEntity getById(Integer id);
+    PaymentEntity updateStatusById(Integer id, Status paymentStatus);
+    void deleteById(Integer id);
 }
